+++ python/router/__init__.py
@@ -3,7 +3,7 @@
 """
 
 from typing import Callable
-from message import Message
+from python_agent_utils.messages.message import Message
 
 
 class BaseRouter:
