+++ python/router/family_router.py
@@ -3,7 +3,7 @@
 
 import re
 from modules import Module
-from message import Message
+from python_agent_utils.messages.message import Message
 from . import BaseRouter, RouteAlreadyRegisteredException, UnparsableMessageFamilyException
