+++ python_agent_utils/messages/basicmessage.py
@@ -1,0 +1,38 @@
+import datetime
+from .message import Message
+
+
+class BasicMessage:
+    FAMILY_NAME = "basicmessage"
+    VERSION = "1.0"
+    FAMILY = "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/" + FAMILY_NAME + "/" + VERSION + "/"
+
+    MESSAGE = FAMILY + "message"
+
+    @staticmethod
+    def build(content: str) -> Message:
+        sent_time = datetime.datetime.utcnow().replace(tzinfo=datetime.timezone.utc).isoformat(' ')
+        return Message({
+            '@type': BasicMessage.MESSAGE,
+            '~l10n': {'locale': 'en'},
+            'sent_time': sent_time,
+            'content': content
+        })
+
+    @staticmethod
+    def validate(msg: Message):
+        msg.validate(
+            [
+                ('@type', BasicMessage.MESSAGE),
+                '~l10n',
+                'sent_time',
+                'content',
+            ]
+        )
+
+        Message.validate_message(
+            [
+                ('locale', 'en')
+            ],
+            msg['~l10n']
+        )
