+++ test_suite/tests/basicmessage/__init__.py
@@ -1,0 +1,1 @@
+
