+++ test_suite/tests/connection/manual.py
@@ -2,8 +2,8 @@ import pytest
 from test_suite.tests import expect_message, pack, unpack, sign_field, get_verified_data_from_signed_field, \
     expect_silence, check_problem_report
 from indy import did
-from modules.connection import Connection
-from modules.did_doc import DIDDoc
+from python_agent_utils.messages.connection import Connection
+from python_agent_utils.messages.did_doc import DIDDoc
 
 
 expect_message_timeout = 30
